improve import logic
The core of this change is to have module names tied to file names, much like
Python. As an example, importing "hashmap" looks for a file in a CBCVM_PATH
named "hashmap.rbcvm", and the resulting module is called "hashmap".

This has some implications:
- Since the module name is now the file name, there is no more "module header".
- "module" is no longer a keyword.
- File names must now be valid identifiers (if they are to be imported).
- There is no longer a concept of a "script" which is not module.

While at it, I added some logic to avoid reading a module's file if we've
already imported it once.

Closes #18
Closes #6