use linear probing for hashmap
This has huge benefits to memory usage and performance in this case. It's still
a simple implementation (probably simpler than the old one), but it drastically
reduces the number of objects that need to be allocated to store a single object
in the hash map.

Each "bucket" in the hash map is now just a struct with a key and a value.
Previously, every bucket would be an assoclist, which is an arraylist. The
initial size of an arraylist is 16 elements, so every bucket would require
allocating a 16-element array, an arraylist struct, and an array of length 2 to
store the key-value pair for the assoclist.