ensure some functions are always inlined
Presumably due to the size of the interpreter loop, GCC decides to stop inlining
functions at some point. This is not ideal for things like stack_push/pop, or
for cb_bytecode_get (though that one isn't used in the interpreter loop anymore
as of a couple of commits ago).